feat: CLIN-1554 - remove legacy index
