+++ src/main/java/bio/ferlab/clin/es/config/ElasticsearchConfiguration.java
@@ -1,9 +1,8 @@
 package bio.ferlab.clin.es.config;
 
-import bio.ferlab.clin.es.indexer.IndexerTools;
+import bio.ferlab.clin.es.data.ElasticsearchData;
 import bio.ferlab.clin.properties.BioProperties;
 import bio.ferlab.clin.es.ElasticsearchRestClient;
-import bio.ferlab.clin.es.data.ElasticsearchData;
 import bio.ferlab.clin.utils.JsonGenerator;
 import org.apache.http.HttpHost;
 import org.apache.http.impl.nio.client.HttpAsyncClientBuilder;
