--- src/main/java/bio/ferlab/clin/es/data/WithFullText.java
@@ -1,7 +1,0 @@
-package bio.ferlab.clin.es.data;
-
-public interface WithFullText {
-  
-  void applyFullText();
-  
-}
