+++ src/main/java/bio/ferlab/clin/es/data/nanuq/AbstractPrescriptionData.java
@@ -1,6 +1,5 @@
 package bio.ferlab.clin.es.data.nanuq;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
