+++ src/main/java/bio/ferlab/clin/properties/BioProperties.java
@@ -31,7 +31,6 @@ public class BioProperties {
     private final String authClientId;
     private final String authSystemId;
     private final String authClientSecret;
-    private final boolean isNanuqEnabled;
     private final String nanuqEsAnalysesIndex;
     private final String nanuqEsSequencingsIndex;
 
@@ -56,7 +55,6 @@ public class BioProperties {
             @Value("${bio.auth.authorization.client-id}") String authClientId,
             @Value("${bio.auth.authorization.client-secret}") String authClientSecret,
             @Value("${bio.auth.authorization.system-id}") String authSystemId,
-            @Value("${bio.nanuq.enabled}") boolean isNanuqEnabled,
             @Value("${bio.nanuq.analyses-index}") String nanuqEsAnalysesIndex,
             @Value("${bio.nanuq.sequencings-index}") String nanuqEsSequencingsIndex
 
@@ -81,7 +79,6 @@ public class BioProperties {
         this.authClientId = authClientId;
         this.authClientSecret = authClientSecret;
         this.authSystemId = authSystemId;
-        this.isNanuqEnabled = isNanuqEnabled;
         this.nanuqEsAnalysesIndex = nanuqEsAnalysesIndex;
         this.nanuqEsSequencingsIndex = nanuqEsSequencingsIndex;
     }
